Conflicts:
	Makefile.thoni

+++ converter/lmList.c
@@ -397,9 +397,9 @@ static void insert(
       /* escaped character */
       msidx++;
       rb[rbidx++] = mstr[msidx++];
-    } else if (mstr[msidx] == '%' && isdigit(mstr[msidx+1])) {
+    } else if (mstr[msidx] == '%' && isdigit((int)mstr[msidx+1])) {
       /* insertion string */
-      for (msidx++, iptr = 0; isdigit(mstr[msidx]); msidx++)
+      for (msidx++, iptr = 0; isdigit((int)mstr[msidx]); msidx++)
         iptr = iptr * 10 + mstr[msidx] - '0';
 
       if (iptr > 0) {
@@ -558,7 +558,7 @@ static void geterr(
         && (msarr[*last].pos.file == fil) 
         && (msarr[*last].pos.line == line) 
        ; (*last)++)
-      *errflg |= msarr[*last].sev; /* this severity was found */
+    *errflg |= msarr[*last].sev; /* this severity was found */
     (*last)--;
   }
 }
@@ -699,7 +699,7 @@ static void prsrcl(
   
   /* Possibly output source if requested */
   if (inset(errflg, lstsev)) {
-    /* There is a message on this line that we want to show */
+    /* There is a message in this line that we want to show */
     if (inset(liMSG, lsttyp)) {
       if (!pageSkipped)
 	skippage();			/* Skip list to next page */
@@ -708,7 +708,7 @@ static void prsrcl(
       prlin(lbuf, TRUE, FALSE, 0);	/* Error line to follow */
     }
   } else {
-    /* No interesting message on this line, show it anyway ? */
+    /* No interesting message on this line, show it anyway? */
     if (inset(liOK, lsttyp)) {
       if (!pageSkipped)
 	skippage();			/* Skip list to next page */
@@ -939,7 +939,7 @@ static void liOpenOutput(
   else {
     out.file = fopen(out.name, "w");
     if (out.file == NULL) {	/* Couldn't open list file */
-      out.file = stdout;	/* So list on standard output instead */
+      out.file = stdout;	/* So list on standard output */
       out.name = NULL;		/* Remember! */
     }
   }
